input/touchscreen: Merge TOUCHSCREEN_OFFLOAD from Floral
The floral driver is like 6 versions older than the sunfish driver we're currently using(originally launched with 11 Beta), but the floral driver supports offloading the touchscreen for better performance. It relies on the TOUCHSCREEN_HEATMAP code we're already using for Sunfish so lets just remove those lines to match floral.